requirements.txt
requests==2.21.0
Django==2.2.1
django-extensions==2.1.6
django-filter==2.1.0
django-mptt==0.10.0
django-redis==4.10.0
djangorestframework==3.9.3
mysqlclient==1.4.2.post1
pytz==2019.1
redis==3.2.1
urllib3==1.24.3
gunicorn==19.9.0
gevent==1.3.7
